IMPROVE: split code to separated files
