+++ packages/app/features/accounts/sign-in/sign-in-form.tsx
@@ -1,48 +1,19 @@
-import { ActivityIndicator, Pressable, Text, View } from 'app/design-system'
-import { Platform } from 'react-native'
-import BackgroundImage from 'app/ui/background-image'
-import WalletconnectBtn from 'app/ui/walletconnect-btn'
-import KeyboardAvoidingView from 'app/design-system/keyboard-avoiding-view'
-import { Separator } from 'app/features/accounts/or-separator'
-import StyledTextInput from 'app/features/accounts/styled-text-input'
-import { Formik, FormikProps } from 'formik'
-import * as Yup from 'yup'
-import { SchemaOf } from 'yup'
 import { useMutation } from '@apollo/client'
 import { REQUEST_TOKEN } from 'app/api/user'
-import { useState } from 'react'
-import { SignInParam, useSignInParam } from 'app/hooks/use-sign-in-param'
-import { useSignIn } from 'app/hooks/use-sign-in'
 import { SignInForm as FormData } from 'app/types'
-import { openEmail } from 'app/utils/email'
-
-export function SignIn() {
-  const signInParam = useSignInParam()
-  const [emailSend, setEmailSend] = useState(false)
-
-  return (
-    <KeyboardAvoidingView
-      behavior={Platform.OS === 'ios' ? 'padding' : undefined}
-      className="absolute inset-0 flex items-center justify-center"
-    >
-      <BackgroundImage />
-
-      {emailSend ? (
-        <OpenEmailView />
-      ) : signInParam ? (
-        <AuthenticationView signInParam={signInParam} />
-      ) : (
-        <SignInForm onEmailSend={() => setEmailSend(true)} />
-      )}
-    </KeyboardAvoidingView>
-  )
-}
+import * as Yup from 'yup'
+import { SchemaOf } from 'yup'
+import { ActivityIndicator, Pressable, Text, View } from 'app/design-system'
+import WalletconnectBtn from 'app/ui/walletconnect-btn'
+import { Separator } from 'app/features/accounts/or-separator'
+import { Formik, FormikProps } from 'formik'
+import StyledTextInput from 'app/features/accounts/styled-text-input'
 
 type SignInFormProps = {
   onEmailSend: () => void
 }
 
-function SignInForm({ onEmailSend }: SignInFormProps) {
+export function SignInForm({ onEmailSend }: SignInFormProps) {
   const [requestToken, { loading, error }] = useMutation(REQUEST_TOKEN, {
     onCompleted: onEmailSend,
   })
@@ -118,36 +89,6 @@ function SignInForm({ onEmailSend }: SignInFormProps) {
           </View>
         )}
       </Formik>
-    </View>
-  )
-}
-
-function OpenEmailView() {
-  return (
-    <View>
-      <Text className="flex flex-row items-center text-sm w-full h-12 rounded-lg p-2 bg-gray-700/20">
-        We send you an email to access your account!
-      </Text>
-      <Pressable onPress={() => openEmail()} className="btn mt-4">
-        <Text className="tracking-0.5">Open Email</Text>
-      </Pressable>
-    </View>
-  )
-}
-
-function AuthenticationView({ signInParam }: { signInParam: SignInParam }) {
-  const { user, error } = useSignIn(signInParam)
-
-  return (
-    <View className="w-72">
-      <Text className="text-center align-center text-sm w-full rounded-lg p-3 bg-gray-700/20">
-        {user ? 'Success' : 'Authentication...'}
-      </Text>
-      {error && (
-        <Text className="w-full text-center text-sm text-[#d9544f] mt-4 min-h-5">
-          {error.message}
-        </Text>
-      )}
     </View>
   )
 }
